@@ -1,4 +1,4 @@
-// swift-tools-version: 5.5
+// swift-tools-version: 6.0
 
 // WARNING:
 // This file is automatically generated.
@@ -10,18 +10,17 @@ import AppleProductTypes
 let package = Package(
     name: "Gravity",
     platforms: [
-        .iOS("15.2")
+        .iOS("16.0")
     ],
     products: [
         .iOSApplication(
             name: "Gravity",
             targets: ["AppModule"],
-            bundleIdentifier: "gravity.com",
-            teamIdentifier: "PC43V2F4N4",
-            displayVersion: "1.1",
-            bundleVersion: "5",
-            iconAssetName: "AppIcon",
-            accentColorAssetName: "AccentColor",
+            bundleIdentifier: "com.sachin.Gravity",
+            displayVersion: "1.0",
+            bundleVersion: "1",
+            appIcon: .asset("AppIcon"),
+            accentColor: .asset("AccentColor"),
             supportedDeviceFamilies: [
                 .pad,
                 .phone
@@ -37,10 +36,8 @@ let package = Package(
     targets: [
         .executableTarget(
             name: "AppModule",
-            path: ".",
-            resources: [
-                .process("Resources")
-            ]
+            path: "."
         )
-    ]
+    ],
+    swiftLanguageVersions: [.version("6")]
 )
