@@ -2,7 +2,7 @@
 import SwiftUI
 
 public struct SOCManager {
-    @available(iOSApplicationExtension, unavailable)
+    @MainActor @available(iOSApplicationExtension, unavailable)
     public static func present<Content:View>(isPresented: Binding<Bool>, onDismiss: (() -> Void)? = nil, options: SOCOptions = SOCOptions(), style: UIUserInterfaceStyle = .unspecified, @ViewBuilder content: @escaping () -> Content) {
         let rootCard = SlideOverCard(isPresented: isPresented, onDismiss: {
             dismiss(isPresented: isPresented)
