@@ -55,33 +55,34 @@ extension View {
 
 struct AnimationCompletionObserverModifier<Value>: AnimatableModifier where Value: VectorArithmetic {
 
-    var animatableData: Value {
+    nonisolated(unsafe) var animatableData: Value {
         didSet {
             notifyCompletionIfFinished()
         }
     }
 
-    private var targetValue: Value
-
+    nonisolated(unsafe) private var targetValue: Value
     private var completion: () -> Void
 
     init(observedValue: Value, completion: @escaping () -> Void) {
         self.completion = completion
         self.animatableData = observedValue
-        targetValue = observedValue
+        self.targetValue = observedValue
     }
 
-    private func notifyCompletionIfFinished() {
-        guard animatableData == targetValue else { return }
-        DispatchQueue.main.async {
-            self.completion()
+    nonisolated(unsafe) private func notifyCompletionIfFinished() {
+        if animatableData == targetValue {
+            DispatchQueue.main.async {
+                self.completion()
+            }
         }
     }
 
     func body(content: Content) -> some View {
-        return content
+        content
     }
 }
+
 
 extension View {
     func onAnimationCompleted<Value: VectorArithmetic>(for value: Value, completion: @escaping () -> Void) -> ModifiedContent<Self, AnimationCompletionObserverModifier<Value>> {
